requirements.txt
Django>=5.0.1,<5.1
djangorestframework>=3.14.0,<3.15
psycopg2==2.9.9
drf-spectacular>=0.27.1,<0.28
Pillow>=10.2.0,<10.3.0